+++ src/MogwaiNano/MogwaiNano/Engine/UdpServer.cs
@@ -1,7 +1,19 @@
-﻿using nanoFramework.Json;
+﻿// Copyright 2026 Stéphane Sibué
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+using nanoFramework.Json;
 using nanoFramework.Runtime.Native;
-using System;
-using System.Diagnostics;
 using System.Net;
 using System.Net.NetworkInformation;
 using System.Net.Sockets;
@@ -15,7 +27,7 @@ namespace MogwaiNano.Engine
         private Thread _udpThread;
         private UdpClient _udpListener;
         private bool _shuttingDown = false;
-        
+
         public void StartUdpServer()
         {
             _udpThread = new Thread(UdpListenLoop);
@@ -76,7 +88,7 @@ namespace MogwaiNano.Engine
         private void SendDiscoveryResponse(UdpClient udpListener, IPEndPoint remoteEndPoint)
         {
             var response = new ServerMessage(
-                AppGlobal.DEVICE_NAME,
+                AppGlobal.NanoParameters.Name,
                 "I AM HERE",
                 AppGlobal.MogwaiNanoEngine.Version.ToString(),
                 AppGlobal.Session.ToString(),